move get org functions into wikidata module
