+++ tests/wikidata/test_extract.py
@@ -3,7 +3,7 @@ from mex.common.wikidata.models.organization import WikidataOrganization
 from mex.common.wikidata.transform import (
     transform_wikidata_organization_to_extracted_organization,
 )
-from mex.pipeline.wikidata import (
+from mex.wikidata.extract import (
     get_organization_merged_id_by_query,
     get_organization_merged_id_by_query_with_transform_and_load,
 )
@@ -23,6 +23,7 @@ def test_get_organization_merged_id_by_query(
     extracted_organization = transform_wikidata_organization_to_extracted_organization(
         wikidata_organization, wikidata_primary_source
     )
+    assert extracted_organization is not None
     returned = get_organization_merged_id_by_query(
         {"foo": wikidata_organization}, wikidata_primary_source
     )
